pyproject.toml
[project]
name = "ai-chatbot"
version = "0.1.0"
description = "An AI chatbot project"
requires-python = ">=3.9"
dependencies = [
    "openai",
    "streamlit",
    "python-dotenv",
    "notebook",
    "ipywidgets"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ai_chatbot"]
